+++ src/interfaces/OrderManager/IOrderManagerErrors.sol
@@ -7,5 +7,5 @@ interface IOrderManagerErrors {
     /// @notice Indicates that the provided tick is not a multiple of the tick spacing or outside of the limits.
     error OrderManagerInvalidTick(int24 tick);
     /// @notice Indicates that the current user tried to withdraw more assets than their deposits.
-    error OrderManagerInsufficientFunds(int24 tick, address user, uint232 amountInTick, uint232 amountToWithdraw);
+    error OrderManagerInsufficientFunds(int24 tick, address user, uint256 amountInTick, uint256 amountToWithdraw);
 }
